Bug fixing and renaming
The to and from links of the "rejected" requests were not added to the relevant links list. Now it is added.

Furthermore, the name "rejected" requests is not accurate here. It is rather "pending" requests, as they may be re-considered in the next rounds, until it is formally rejected at the "submission time".